+++ api/src/signals/integrations/sigmax/pdf.py
@@ -1,10 +1,15 @@
 """
 This module contains the PDF generation code used for Sigmax.
 """
-import base64
+import os
 import logging
+import base64
+import warnings
+
 
 import weasyprint
+from dateutil.parser import parse
+
 from django.template.loader import render_to_string
 
 # Because weasyprint can produce a lot of warnings (unsupported
@@ -15,24 +20,24 @@ logging.getLogger('weasyprint').setLevel(100)
 
 
 def _render_html(signal: Signal):
-    dt = signal.created_at
+    dt = parse(signal['created_at'])
 
-    return render_to_string('sigmax/pdf_template.html', context={
+    return render_to_string('pdf_template.html', context={
         'signal': signal,
         'datum': dt.strftime('%Y %m %d'),
         'tijdstip': dt.strftime('%H:%M:%S'),
         'hoofdrubriek': signal.category.main,
         'subrubriek': signal.category.sub,
-        'omschrijving': signal.text,
+        'omschrijving': signal['text'],
         'stadsdeel': '',  # TODO, extract from gebieden API?
-        'adres': signal.location.address_text,
-        'bron': signal.source,
+        'adres': signal['location']['address_text'],
+        'bron': signal['source'],
         'email': 'placeholder@example.com',
         'telefoonnummer': 'nvt',
     })
 
 
-def _generate_pdf(signal: Signal):
+def _generate_pdf(signal):
     """
     Generate PDF to send to VoegZaakdocumentToe_Lk01 (returns base64 encoded data).
     """
