+++ api/src/signals/integrations/sigmax/pdf_template.html
@@ -1,0 +1,56 @@
+<!DOCTYPE html>
+<html>
+<head lang="en">
+    <meta charset="UTF-8">
+    <title>Melding MORA-{{ signal.id }} </title>
+	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
+</head>
+</head>
+<body>
+
+<h2>Melding MORA-{{ signal.id }} </h2>
+<table style="width:100%" class="table">
+  <tr>
+    <td>Datum</td>
+    <td>{{ datum }}</td> 
+  </tr>
+  <tr>
+    <td>Tijdstip</td>
+    <td>{{ tijdstip }}</td> 
+  </tr>
+  <tr>
+    <td>Hoofdrubriek</td>
+    <td>{{ hoofdrubriek }}</td> 
+  </tr>
+  <tr>
+    <td>Subrubriek</td>
+    <td>{{ subrubriek }}</td> 
+  </tr>
+  <tr>
+    <td>Omschrijving</td>
+    <td>{{ omschrijving }} </td> 
+  </tr>
+  <tr>
+    <td>Stadsdeel</td>
+    <td>{{ stadsdeel }}</td> 
+  </tr>
+  <tr>
+    <td>Adres</td>
+    <td>{{ adres }}</td> 
+  </tr>
+  <tr>
+    <td>Bron</td>
+    <td>{{ bron }}</td> 
+  </tr>
+  <tr>
+    <td>Email</td>
+    <td>{{ email }}</td> 
+  </tr>
+  <tr>
+    <td>Telefoonnummer</td>
+    <td>{{ telefoonnummer }}</td> 
+  </tr>
+</table>
+
+</body>
+</html>
